src/main/java/lottoauto/service/LottoGame.java
package lottoauto.service;

import lottoauto.domain.Lotto;

/**
 * - 금액을 입력 받아 로또 게임을 하는 클래스를 만든다.
 * - 1000원 금액에 맞는 로또 갯수만큼 로또 클래스를 생성한다.
 * - 당첨 로또를 생성한다.
 * - 당첨 통계를 출력한다.
 */
public class LottoGame {
    private final int money;
    private final int lottoCount;

    public LottoGame(int money) {
        if(money < Lotto.PRICE) {
            throw new IllegalArgumentException("로또를 살 수 없습니다.");
        }
        this.money = money;
        this.lottoCount = money / Lotto.PRICE;
    }

    public int getLottoCount() {
        return lottoCount;
    }

}
